# src/future-be/src/swagger.json
{
	"components": {
...
		"schemas": {
...
			"FlattenMaps_T_": {
				"properties": {},
				"type": "object"
			},
			"IProductModel": {
				"$ref": "#/components/schemas/FlattenMaps_T_"
			},
			"IOrderItem": {
				"properties": {
					"product": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/IProductModel"
							}
						]
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"product",
					"price",
					"quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateOrder": {
				"properties": {
					"orderItems": {
						"items": {
							"$ref": "#/components/schemas/IOrderItem"
						},
						"type": "array"
					},
					"address": {
						"type": "string"
					},
					"paymentMethod": {
						"type": "string"
					}
				},
				"required": [
					"orderItems",
					"address",
					"paymentMethod"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOrderHistoryRes": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"shortId": {
						"type": "string"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string"
					},
					"firstProduct": {
						"properties": {
							"price": {
								"type": "number",
								"format": "double"
							},
							"quantity": {
								"type": "number",
								"format": "double"
							},
							"thumbnail": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"price",
							"quantity",
							"thumbnail",
							"name",
							"_id"
						],
						"type": "object"
					},
					"orderItemsLength": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"shortId",
					"total",
					"createdAt",
					"firstProduct",
					"orderItemsLength"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrderStatus": {
				"enum": [
					"pending",
					"delivering",
					"completed"
				],
				"type": "string"
			},
			"IAddress": {
				"properties": {
					"province": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"ward": {
						"type": "string"
					},
					"specificAddress": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"receiver": {
						"type": "string"
					},
					"default": {
						"type": "boolean"
					},
					"deleted": {
						"type": "boolean"
					}
				},
				"required": [
					"province",
					"district",
					"ward",
					"specificAddress",
					"phone",
					"receiver",
					"default",
					"deleted"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserModel": {
				"$ref": "#/components/schemas/FlattenMaps_T_"
			},
			"IAllOrders": {
				"properties": {
					"shortId": {
						"type": "string"
					},
					"address": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/IAddress"
							}
						]
					},
					"userName": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/IUserModel"
							}
						]
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"dateCreated": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/OrderStatus"
					}
				},
				"required": [
					"shortId",
					"address",
					"userName",
					"total",
					"dateCreated",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOrderRes": {
				"properties": {
					"allOrders": {
						"items": {
							"$ref": "#/components/schemas/IAllOrders"
						},
						"type": "array"
					},
					"numOfProds": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"allOrders",
					"numOfProds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRevenueValue": {
				"properties": {
					"value": {
						"type": "number",
						"format": "double"
					},
					"label": {
						"type": "string"
					}
				},
				"required": [
					"value",
					"label"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "future-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/orders": {
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateOrder"
							}
						}
					}
				}
			}
		},
		"/orders/history": {
			"get": {
				"operationId": "GetOrderHistoryfollowStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/IOrderHistoryRes"
													},
													"type": "array"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/OrderStatus"
						}
					}
				]
			}
		},
		"/orders/pagination": {
			"get": {
				"operationId": "GetAllOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/IOrderRes"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/orders/order-follow-date/pagination": {
			"get": {
				"operationId": "GetOrdersFollowDate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/IOrderRes"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/orders/update-status": {
			"put": {
				"operationId": "UpdateStatusOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "order-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/OrderStatus"
						}
					}
				]
			}
		},
		"/orders/order-by-id": {
			"get": {
				"operationId": "GetDetailOrderById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/IOrderItem"
													},
													"type": "array"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "order-id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orders/revenue": {
			"get": {
				"operationId": "GetRevenueFollowTime",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/IRevenueValue"
													},
													"type": "array"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "timeReport",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orders/revenue-current-year": {
			"get": {
				"operationId": "GetRevenueOfCurrentYear",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "number",
													"format": "double"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}